UPDATE: Bugfix for map votes
Fixed issue where map lists were not updating. This was because the map votes were being deregistered prior to the map list updates and throwing an exception. The exception was because deregistration was not scoped to the current map group, but instead all maps from every map group.

Thanks @mavproductions!